refactor(runtime): split evidence ledger types
